requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-forms-bootstrap==3.1.0
django-paypal==0.3.6
djangorestframework==3.7.3
djangorestframework-jwt==1.8.0
idna==2.6
PyJWT==1.4.2
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
